Refactor the import legacy persist logic
We have recently introduced functionality to import data from NALD using for a licence and companies for that particular licence.

This change will break out the persist functionality into logical groups of functionality.
